% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSimData.R
\name{GenSimData}
\alias{GenSimData}
\title{Generate simulated data}
\usage{
GenSimData(
  k,
  ni = 100,
  xdim = 5,
  xpdf = "norm",
  sd = sqrt(0.01),
  tau_fn = "X1*(X1>0) -4*(k\%\%2==0)+X1*(k\%\%2==0)",
  m_fn = "0"
)
}
\arguments{
\item{k}{Site index.}

\item{ni}{Sample size within a site.}

\item{xdim}{Dimension of covariate `X`.}

\item{xpdf}{Distribution of covariate `X`.}

\item{sd}{Standard deviation of error term in simulating outcome.}

\item{tau_fn}{Treatment effect function.}

\item{m_fn}{Mean effect function.}
}
\value{
A simulated data set.
}
\description{
Generate simulated data
}
\examples{
GenSimData(k=1)

}
