% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSimData.R
\name{GenSimX}
\alias{GenSimX}
\title{Simulate X covariates}
\usage{
GenSimX(ni = 100, xdim = 5, xpdf = "norm")
}
\arguments{
\item{ni}{Sample size within a site.}

\item{xdim}{Dimension of covariate `X`.}

\item{xpdf}{Distribution of covariate `X`.}
}
\value{
Covariate matrix `X`.
}
\description{
Simulate X covariates
}
\examples{
GenSimX()

}
