% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetHonestRcpp}
\alias{GetHonestRcpp}
\title{Get honest predictions (C++)}
\source{
\url{https://github.com/okasag/orf/blob/master/orf/src/get_honest_rcpp.cpp}
}
\usage{
GetHonestRcpp(x, y, z, w)
}
\arguments{
\item{x}{unique_leaves (List)[ntree]}

\item{y}{honest_y (NumericVector)[nrow]}

\item{z}{honest_leaves (NumericMatrix)[nrow, ntree]}

\item{w}{train_leaves (NumericMatrix)[nrow, ntree]}
}
\description{
Computes honest predictions (fitted values) from the random forest
for the train and honest sample based on the honest training sample.
}
\keyword{internal}
