% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotForestImp.R
\name{PlotForestImp}
\alias{PlotForestImp}
\title{Importance plot of an ensemble forest}
\usage{
PlotForestImp(myfit)
}
\arguments{
\item{myfit}{A fitted ensemble forest.}
}
\value{
A ggplot2 object and a table of relative importance of each covariate.
}
\description{
Importance plot of an ensemble forest
}
\examples{
data(SimDataLst)
K <- length(SimDataLst)
covars <- grep("^X", names(SimDataLst[[1]]), value=TRUE)
fit_lst <- list()
for (k in 1:K) {
    tmpdf <- SimDataLst[[k]]
    # use your estimator of interest
    fit_lst[[k]] <- grf::causal_forest(X=as.matrix(tmpdf[, covars, with=FALSE]),
                                       Y=tmpdf$Y, W=tmpdf$Z)
}

coord_id <- 1
coord_df <- SimDataLst[[coord_id]]
aug_df <- GenAugData(coord_id, coord_df, fit_lst, covars)

## Treat each site as a distinct factor
myfit <- EnsemForest(coord_id, aug_df, "site", covars, importance="impurity")$myfit
PlotForestImp(myfit)

## Mean encoding as surrogate for site index
myfit <- EnsemForest(coord_id, aug_df, "site", covars, is_encode=TRUE)$myfit
PlotForestImp(myfit)


}
