% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PredHonestRcpp}
\alias{PredHonestRcpp}
\title{Predict honest predictions (C++)}
\source{
\url{https://github.com/okasag/orf/blob/master/orf/src/pred_honest_rcpp.cpp}
}
\usage{
PredHonestRcpp(x, y, z, w)
}
\arguments{
\item{x}{unique_leaves (List)[ntree]}

\item{y}{honest_y (NumericVector)[nrow]}

\item{z}{honest_leaves (NumericMatrix)[nrow, ntree]}

\item{w}{test_leaves (NumericMatrix)[nrow, ntree]}
}
\description{
Computes honest predictions from the random forest for
a test sample based on the honest training sample.
}
\keyword{internal}
